+++ pyArango/validation.py
@@ -15,8 +15,8 @@ class Validator(object) :
 
 class NotNull(Validator) :
     """Checks that the Field has a non null value"""
-    def validate(self, value) :
-        if value is None or value == "" :
+    def validate(self, value, zero=True, emptyString=True) :
+        if value is None or (value == 0 is zero) or (value == "" and emptyString) :
             raise ValidationError("Field can't have a null value: '%s'" % value)
         return True
 
@@ -29,6 +29,60 @@ class Email(Validator) :
             raise ValidationError("The email address: %s is invalid" % value)
         return True
 
+class Numeric(Validator) :
+    """checks if the value is numerical"""
+    def validate(self, value) :
+        try :
+            float(value)
+        except :
+            raise ValidationError("%s is not valid numerical value" % value)
+        return True
+
+class Int(Validator) :
+    """The value must be an integer"""
+    def validate(self, value) :
+        if not isinstance(value, int) :
+            raise ValidationError("%s is not a valid integer" % value)
+        return True
+
+class Bool(Validator) :
+    """The value must be a boolean"""
+    def validate(self, value) :
+        if not isinstance(value, bool) :
+            raise ValidationError("%s is not a valid boolean" % value)
+        return True
+
+class String(Validator) :
+    """The value must be a string or unicode"""
+    def validate(self, value) :
+        if not isinstance(value, str) and not isinstance(value, unicode) :
+            raise ValidationError("%s is not a valid string" % value)
+        return True
+
+class Enumeration(Validator) :
+    """The value must be in the allowed ones"""
+    def __init__(self, allowed) :
+        self.allowed = set(allowed)
+  
+    def validate(self, value) :
+        if value not in self.allowed :
+            raise ValidationError("%s is not among the allowed values %s" % (value, self.allowed))
+        return True
+
+class Range(Validator) :
+    """The value must une [lower, upper] range"""
+    def __init__(self, lower, upper) :
+        self.lower = lower
+        self.upper = upper
+
+    def validate(self, value) :
+        if value < self.lower or value > self.upper :
+            raise ValidationError("%s is not in [%s, %s]" % (value, self.lower, self.upper))
+    
+    def __str__(self) :
+        return "%s[%s, %s]" % (self.__class__.__name__, self.minLen, self.maxLen)
+
+
 class Length(Validator) :
     """validates that the value length is between given bounds"""
     def __init__(self, minLen, maxLen) :
@@ -36,12 +90,14 @@ class Length(Validator) :
         self.maxLen = maxLen
 
     def validate(self, value) :
-        try:
-            if self.minLen <= len(value) and len(value) <= self.maxLen :
-                return True
-            raise ValidationError("Field must have a length in ['%s';'%s'] got: '%s'" % (self.minLen, self.maxLen, len(value)))
-        except Exception as e:
-            raise ValidationError("Invalid value: '%s' has no length" % value)
-
+        try :
+            length = len(value)
+        except :
+            raise ValidationError("Field '%s' of type '%s' has no length" % (value, type(value)))
+            
+        if self.minLen <= len(value) and len(value) <= self.maxLen :
+            return True
+        raise ValidationError("Field must have a length in ['%s';'%s'] got: '%s'" % (self.minLen, self.maxLen, len(value)))
+    
     def __str__(self) :
         return "%s[%s, %s]" % (self.__class__.__name__, self.minLen, self.maxLen)