+++ CHANGELOG.rst
@@ -1,8 +1,32 @@
+<<<<<<< HEAD
 1.2.9
 =====
 
 * Added bulk import to connection
 * Added binvars to explain
+=======
+1.3.2
+=====
+
+* Validation bug fixes
+* New Numeric, Int, Bool, String, Enumeration, Range validators
+* Fields can have default values
+* When creationg a new document, Collection will serve one populated with defaults
+* stastd support thx to: @dothebart
+
+1.3.1
+=====
+
+* Will die gracefully if server response is empty
+* getStore and getPatches shorthands added to Document
+
+1.3.0
+=====
+
+* Fixed nested store patch update
+* REFACT: New DocumentStore class for taking care of storing document in a hierarchy of stores (nested objects) and validate them
+* Minor bug fixes
+>>>>>>> f314570290f59c1f23aa74d5efd1324c2fcf3993
 
 1.2.8
 =====